moved exception handler to gather all the exception handler in to one package
rename class ExceptionHandler since it is conflicting with @ExceptionHandler annotation
