--- src/main/java/com/example/myinsta/exception/ExceptionHandler.java
@@ -1,6 +1,0 @@
-package com.example.myinsta.exception;
-
-
-public class ExceptionHandler {
-
-}
